refactor(internal): move internally used components to package internal
