+++ internal/doer.go
@@ -1,4 +1,4 @@
-package api
+package internal
 
 import (
 	"net/http"
